[NPC sell / Buy] -> solved bug when selling jems and equips to NPC (was 0), -> add anti hack for stackables sell to NPC (check before to give
Zuly),
-> change how thei- anthack is handled when selling to NPC (exits 

immediatly from loop),

[Trading]
-> solved bug check when trading Zuly between 2 players (it was 

checking the two amounts against the same player),
-> msg anti hack for Zuly added,
-> check all items to be traded before actually giving them,

[Shop]
-> it seems checks are already done there (for buy / sell),

[Clan]
-> custom clanmark should work now (needs an extra Int (11) field 

in list_clan table: siglogo),

[JRose (thanks zrose)]
-> unlimited gaming,
-> quick fix so the storage doesn't crash anymore when adding item,


[Known Bug]
-> storage bug sometimes, last item is not seen anymore,

[Solved bug]
-> GM and dev skills not learned anymore (rl2171),
-> 2 dealer summons fixs (thanks ??? in forum)

+++ trunk/World Server/WorldServer.h
@@ -255,7 +255,7 @@ class CWorldServer : public CServerSocket
         bool pakGMItemQuest( CPlayer* thisclient, UINT itemid, UINT itemtype, UINT itemamount, UINT itemrefine, UINT itemls, UINT itemstats, UINT itemsocket, char buffer2[200] );
         bool pakGMTeleOtherPlayer( CPlayer *thisclient, char* name, int map, float x, float y );    
         bool pakGMZuly( CPlayer* thisclient, int mode, int amount, char* charname );
-        bool pakGMEventType(CPlayer* thisclient, int npctype, int dialog, int type); //Event
+        bool pakGMEventType(CPlayer* thisclient, int npctype, int dialog, long int type); //Event
     	bool pakGMTele( CPlayer* thisclient, int map, float x, float y );
     	bool pakGMMon( CPlayer* thisclient, int montype, int moncount );    
         bool pakGMZulygive(CPlayer* thisclient, char* name, int zuly);
